update style MovieDetail
