+++ src/pages/movie-detail/MovieDetail.css
@@ -1,0 +1,16 @@
+.detail-section {
+  border: black 1px solid;
+  position: relative;
+  z-index: 1;
+  height: 100vh;
+  background-size: cover;
+}
+
+.detail-content {
+  position: absolute;
+  z-index: 1;
+  bottom: 0;
+  width: 100%;
+  padding: 0 3rem;
+  background: linear-gradient(transparent 55%, black);
+}
