+++ src/routes/routes.jsx
@@ -1,6 +1,6 @@
 import { Route } from "react-router-dom";
 import { Home } from "../pages/home/Home";
-import { MovieDetail } from "../pages/MovieDetail";
+import { MovieDetail } from "../pages/movie-detail/MovieDetail";
 import { Genre } from "../pages/genre/Genre";
 import { SimilarMovie } from "../pages/SimilarMovie";
